feat(binary): option to bind binary directory
Sometimes, in binary mode, host binary directory must be bound instead
of binary path itself. This allows to access contents of binary's directory
within the container. For example, this can useful when the binary
is linked against a shared library located in the same or descendant directory.
Option to do that has been added. It is disabled by default, preserving
old behaviour.